AI-ZK-Agents/scripts/requirements.txt
flask
python-dotenv
pdfplumber
openai
PyGithub
GitPython
nbformat
requests